+++ modules/client/examples/01-tokenVoting-client/06-can-vote.ts
@@ -18,10 +18,9 @@ const contextPlugin: ContextPlugin = ContextPlugin.fromContext(context);
 const client = new TokenVotingClient(contextPlugin);
 
 const voteParams: ICanVoteParams = {
+  proposalId: "0x1234567890123456789012345678901234567890_0x0",
   address: "0x1234567890123456789012345678901234567890",
-  proposalId: "0x1234567890123456789012345678901234567890_0x1",
-  pluginAddress: "0x1234567890123456789012345678901234567890",
-  vote: VoteValues.YES
+  vote: VoteValues.YES,
 };
 
 const canVote = await client.methods.canVote(voteParams);