+++ modules/client/examples/02-addresslistVoting-client/05-vote-proposal.ts
@@ -19,8 +19,7 @@ const contextPlugin: ContextPlugin = ContextPlugin.fromContext(context);
 const client = new AddresslistVotingClient(contextPlugin);
 
 const voteParams: IVoteProposalParams = {
-  pluginAddress: "0x1234567890123456789012345678901234567890",
-  proposalId: "0x1234567890123456789012345678901234567890",
+  proposalId: "0x1234567890123456789012345678901234567890_0x0",
   vote: VoteValues.YES,
 };