+++ modules/client/examples/02-addresslistVoting-client/06-can-vote.ts
@@ -19,9 +19,8 @@ const client = new AddresslistVotingClient(contextPlugin);
 
 const voteParams: ICanVoteParams = {
   address: "0x1234567890123456789012345678901234567890",
-  proposalId: "0x1234567890123456789012345678901234567890_0x1",
-  pluginAddress: "0x1234567890123456789012345678901234567890",
-  vote: VoteValues.YES
+  proposalId: "0x1234567890123456789012345678901234567890_0x0",
+  vote: VoteValues.YES,
 };
 
 const canVote = await client.methods.canVote(voteParams);