+++ modules/client/examples/02-addresslistVoting-client/08-get-proposal.ts
@@ -2,8 +2,8 @@
 ### Loading the a proposal by proposalId (address list plugin)
 */
 import {
-  AddresslistVotingProposal,
   AddresslistVotingClient,
+  AddresslistVotingProposal,
   Context,
   ContextPlugin,
 } from "@aragon/sdk-client";
@@ -16,15 +16,16 @@ const contextPlugin: ContextPlugin = ContextPlugin.fromContext(context);
 // Create an AddressList client
 const client = new AddresslistVotingClient(contextPlugin);
 
-const proposalId = "0x12345...";
+const proposalId = "0x1234567890123456789012345678901234567890_0x0";
 
-const proposal: AddresslistVotingProposal | null = await client.methods.getProposal(
-  proposalId,
-);
+const proposal: AddresslistVotingProposal | null = await client.methods
+  .getProposal(
+    proposalId,
+  );
 console.log(proposal);
 /*
 {
-  id: "0x12345...",
+  id: "0x1234567890123456789012345678901234567890_0x0",
   dao: {
     address: "0x1234567890123456789012345678901234567890",
     name: "Cool DAO"