+++ modules/client/examples/06-multisig-client/07-can-execute.ts
@@ -2,7 +2,6 @@
 ### Checking if user can approve in a multisig plugin
 */
 import {
-  CanExecuteParams,
   Context,
   ContextPlugin,
   MultisigClient,
@@ -15,11 +14,7 @@ const context: Context = new Context(contextParams);
 const contextPlugin: ContextPlugin = ContextPlugin.fromContext(context);
 // Create an multisig client
 const client = new MultisigClient(contextPlugin);
-const canExecuteParams: CanExecuteParams = {
-  pluginAddress: "0x1234567890123456789012345678901234567890",
-  proposalId: 0,
-};
-const canExecute = await client.methods.canExecute(canExecuteParams);
+const canExecute = await client.methods.canExecute("0x1234567890123456789012345678901234567890_0x0");
 console.log(canExecute);
 /*
 true