+++ modules/client/src/tokenVoting/internal/utils.ts
@@ -182,21 +182,21 @@ export function tokenVotingInitParamsToContract(
 function parseToken(
   subgraphToken: SubgraphErc20Token | SubgraphErc721Token,
 ): Erc20TokenDetails | Erc721TokenDetails | null {
-  let token: Erc721TokenDetails| Erc20TokenDetails | null = null;
+  let token: Erc721TokenDetails | Erc20TokenDetails | null = null;
   if (subgraphToken.__typename === SubgraphContractType.ERC20) {
     token = {
       address: subgraphToken.id,
       symbol: subgraphToken.symbol,
       name: subgraphToken.name,
       decimals: subgraphToken.decimals,
-      type: TokenType.ERC20
+      type: TokenType.ERC20,
     };
   } else if (subgraphToken.__typename === SubgraphContractType.ERC721) {
     token = {
       address: subgraphToken.id,
       symbol: subgraphToken.symbol,
       name: subgraphToken.name,
-      type: TokenType.ERC721
+      type: TokenType.ERC721,
     };
   }
   return token;