+++ modules/client/test/integration/multisig-client/estimation.test.ts
@@ -77,8 +77,7 @@ describe("Client Multisig", () => {
       const client = new MultisigClient(ctxPlugin);
 
       const approveParams: ApproveMultisigProposalParams = {
-        proposalId: 0,
-        pluginAddress: "0x1234567890123456789012345678901234567890",
+        proposalId: "0x1234567890123456789012345678901234567890_0x0",
         tryExecution: true,
       };
 
@@ -96,10 +95,7 @@ describe("Client Multisig", () => {
       const ctxPlugin = ContextPlugin.fromContext(ctx);
       const client = new MultisigClient(ctxPlugin);
       const estimation = await client.estimation.executeProposal(
-        {
-          pluginAddress: "0x1234567890123456789012345678901234567890",
-          proposalId: 0,
-        },
+        "0x1234567890123456789012345678901234567890_0x0",
       );
 
       expect(typeof estimation).toEqual("object");