+++ modules/client/test/integration/multisig-client/methods.test.ts
@@ -34,7 +34,7 @@ import {
 } from "../constants";
 import { Server } from "ganache";
 // import { advanceBlocks } from "../../helpers/advance-blocks";
-import { CanExecuteParams, ExecuteProposalStep } from "../../../src";
+import { ExecuteProposalStep } from "../../../src";
 import { buildMultisigDAO } from "../../helpers/build-daos";
 import { mineBlock, restoreBlockTime } from "../../helpers/block-times";
 import { JsonRpcProvider } from "@ethersproject/providers";
@@ -76,7 +76,7 @@ describe("Client Multisig", () => {
   async function buildProposal(
     pluginAddress: string,
     multisigClient: MultisigClient,
-  ): Promise<number> {
+  ): Promise<string> {
     // generate actions
     const action = multisigClient.encoding.updateMultisigVotingSettings(
       {
@@ -124,12 +124,11 @@ describe("Client Multisig", () => {
           expect(step.txHash).toMatch(/^0x[A-Fa-f0-9]{64}$/i);
           break;
         case ProposalCreationSteps.DONE:
-          expect(typeof step.proposalId).toBe("number");
+          expect(typeof step.proposalId).toBe("string");
+          expect(step.proposalId).toMatch(
+            /^0x[A-Fa-f0-9]{40}_0x[A-Fa-f0-9]{1,64}$/,
+          );
           return step.proposalId;
-          // TODO
-        // update with new proposal id when contracts are ready
-        // expect(typeof step.proposalId).toBe("string");
-        // expect(step.proposalId).toMatch(/^0x[A-Fa-f0-9]{64}$/i);
         default:
           throw new Error(
             "Unexpected proposal creation step: " +
@@ -140,13 +139,11 @@ describe("Client Multisig", () => {
     throw new Error();
   }
   async function approveProposal(
-    proposalId: number,
-    pluginAddress: string,
+    proposalId: string,
     client: MultisigClient,
   ) {
     const approveParams: ApproveMultisigProposalParams = {
       proposalId,
-      pluginAddress,
       tryExecution: false,
     };
     for await (const step of client.methods.approveProposal(approveParams)) {
@@ -191,7 +188,6 @@ describe("Client Multisig", () => {
       const canApproveParams: CanApproveParams = {
         proposalId,
         addressOrEns: TEST_WALLET_ADDRESS,
-        pluginAddress,
       };
       // positive
       let canApprove = await client.methods.canApprove(canApproveParams);
@@ -216,7 +212,7 @@ describe("Client Multisig", () => {
       const { plugin: pluginAddress } = await buildDao();
 
       const proposalId = await buildProposal(pluginAddress, client);
-      await approveProposal(proposalId, pluginAddress, client);
+      await approveProposal(proposalId, client);
     });
   });
 
@@ -229,18 +225,14 @@ describe("Client Multisig", () => {
       const { plugin: pluginAddress } = await buildDao();
 
       const proposalId = await buildProposal(pluginAddress, client);
-      const canExecuteParams: CanExecuteParams = {
-        proposalId,
-        pluginAddress,
-      };
-      let canExecute = await client.methods.canExecute(canExecuteParams);
+      let canExecute = await client.methods.canExecute(proposalId);
       expect(typeof canExecute).toBe("boolean");
       expect(canExecute).toBe(false);
 
       // now approve
-      await approveProposal(proposalId, pluginAddress, client);
+      await approveProposal(proposalId, client);
 
-      canExecute = await client.methods.canExecute(canExecuteParams);
+      canExecute = await client.methods.canExecute(proposalId);
       expect(typeof canExecute).toBe("boolean");
       expect(canExecute).toBe(true);
     });
@@ -255,14 +247,11 @@ describe("Client Multisig", () => {
       const { plugin: pluginAddress } = await buildDao();
 
       const proposalId = await buildProposal(pluginAddress, client);
-      await approveProposal(proposalId, pluginAddress, client);
+      await approveProposal(proposalId, client);
 
       for await (
         const step of client.methods.executeProposal(
-          {
-            pluginAddress,
-            proposalId,
-          },
+          proposalId,
         )
       ) {
         switch (step.key) {
@@ -340,7 +329,7 @@ describe("Client Multisig", () => {
       }
       expect(proposal.id).toBe(proposalId);
       expect(typeof proposal.id).toBe("string");
-      expect(proposal.id).toMatch(/^0x[A-Fa-f0-9]{40}_0x[A-Fa-f0-9]{1,}$/i);
+      expect(proposal.id).toMatch(/^0x[A-Fa-f0-9]{40}_0x[A-Fa-f0-9]{1,64}$/i);
       expect(typeof proposal.dao.address).toBe("string");
       expect(proposal.dao.address).toMatch(/^0x[A-Fa-f0-9]{40}$/i);
       expect(typeof proposal.dao.name).toBe("string");