+++ modules/client/test/integration/tokenVoting-client/estimation.test.ts
@@ -5,12 +5,11 @@ declare const describe, it, expect, beforeAll, afterAll;
 import "../../mocks/aragon-sdk-ipfs";
 
 import {
-  TokenVotingClient,
   Context,
   ContextPlugin,
   ICreateProposalParams,
-  IExecuteProposalParams,
   IVoteProposalParams,
+  TokenVotingClient,
   VoteValues,
 } from "../../../src";
 
@@ -57,8 +56,7 @@ describe("Token Voting Client", () => {
       const client = new TokenVotingClient(ctxPlugin);
 
       const voteParams: IVoteProposalParams = {
-        pluginAddress: "0x1234567890123456789012345678901234567890",
-        proposalId: 0,
+        proposalId: "0x1234567890123456789012345678901234567890_0x0",
         vote: VoteValues.YES,
       };
       const estimation = await client.estimation.voteProposal(voteParams);
@@ -74,11 +72,9 @@ describe("Token Voting Client", () => {
       const ctxPlugin = ContextPlugin.fromContext(ctx);
       const client = new TokenVotingClient(ctxPlugin);
 
-      const executeParams: IExecuteProposalParams = {
-        pluginAddress: "0x1234567890123456789012345678901234567890",
-        proposalId: 0,
-      };
-      const estimation = await client.estimation.executeProposal(executeParams);
+      const estimation = await client.estimation.executeProposal(
+        "0x1234567890123456789012345678901234567890_0x0",
+      );
 
       expect(typeof estimation).toEqual("object");
       expect(typeof estimation.average).toEqual("bigint");