+++ modules/client/test/integration/tokenVoting-client/methods.test.ts
@@ -1,17 +1,12 @@
-// @ts-ignore
-declare const describe, it, beforeAll, afterAll, expect, test;
-
 // mocks need to be at the top of the imports
 import { mockedIPFSClient } from "../../mocks/aragon-sdk-ipfs";
 
 import {
-  CanExecuteParams,
   Context,
   ContextPlugin,
   ExecuteProposalStep,
   ICanVoteParams,
   ICreateProposalParams,
-  IExecuteProposalParams,
   IProposalQueryParams,
   IVoteProposalParams,
   ProposalCreationSteps,
@@ -138,10 +133,11 @@ describe("Token Voting Client", () => {
           expect(step.txHash).toMatch(/^0x[A-Fa-f0-9]{64}$/i);
           break;
         case ProposalCreationSteps.DONE:
-          expect(typeof step.proposalId).toBe("number");
+          expect(typeof step.proposalId).toBe("string");
+          expect(step.proposalId).toMatch(
+            /^0x[A-Fa-f0-9]{40}_0x[A-Fa-f0-9]{1,64}$/,
+          );
           return step.proposalId;
-        // TODO fix with new proposalId format
-        // expect(step.proposalId).toMatch(/^0x[A-Fa-f0-9]{64}$/i);
         default:
           throw new Error(
             "Unexpected proposal creation step: " +
@@ -153,13 +149,11 @@ describe("Token Voting Client", () => {
   }
 
   async function voteProposal(
-    pluginAddress: string,
-    proposalId: number,
+    proposalId: string,
     client: TokenVotingClient,
     voteValue: VoteValues = VoteValues.YES,
   ) {
     const voteParams: IVoteProposalParams = {
-      pluginAddress,
       proposalId,
       vote: voteValue,
     };
@@ -195,7 +189,8 @@ describe("Token Voting Client", () => {
           }
 
           const proposalId = await buildProposal(pluginAddress, client);
-          expect(typeof proposalId).toBe("number");
+          expect(typeof proposalId).toBe("string");
+          expect(proposalId).toMatch(/^0x[A-Fa-f0-9]{40}_0x[A-Fa-f0-9]{1,64}$/);
         }
       });
     });
@@ -215,12 +210,12 @@ describe("Token Voting Client", () => {
           }
 
           const proposalId = await buildProposal(pluginAddress, client);
-          expect(typeof proposalId).toBe("number");
+          expect(typeof proposalId).toBe("string");
+          expect(proposalId).toMatch(/^0x[A-Fa-f0-9]{40}_0x[A-Fa-f0-9]{1,64}$/);
 
           const params: ICanVoteParams = {
             address: TEST_WALLET_ADDRESS,
             proposalId,
-            pluginAddress,
             vote: VoteValues.YES,
           };
           const canVote = await client.methods.canVote(params);
@@ -245,10 +240,11 @@ describe("Token Voting Client", () => {
           }
 
           const proposalId = await buildProposal(pluginAddress, client);
-          expect(typeof proposalId).toBe("number");
+          expect(typeof proposalId).toBe("string");
+          expect(proposalId).toMatch(/^0x[A-Fa-f0-9]{40}_0x[A-Fa-f0-9]{1,64}$/);
 
           // Vote
-          await voteProposal(pluginAddress, proposalId, client);
+          await voteProposal(proposalId, client);
         }
       });
 
@@ -266,12 +262,13 @@ describe("Token Voting Client", () => {
         }
 
         const proposalId = await buildProposal(pluginAddress, client);
-        expect(typeof proposalId).toBe("number");
+        expect(typeof proposalId).toBe("string");
+        expect(proposalId).toMatch(/^0x[A-Fa-f0-9]{40}_0x[A-Fa-f0-9]{1,64}$/);
 
         // Vote
-        await voteProposal(pluginAddress, proposalId, client, VoteValues.NO);
+        await voteProposal(proposalId, client, VoteValues.NO);
         await mineBlock(provider);
-        await voteProposal(pluginAddress, proposalId, client, VoteValues.YES);
+        await voteProposal(proposalId, client, VoteValues.YES);
       });
     });
 
@@ -290,22 +287,19 @@ describe("Token Voting Client", () => {
         }
 
         const proposalId = await buildProposal(pluginAddress, client);
-        expect(typeof proposalId).toBe("number");
+        expect(typeof proposalId).toBe("string");
+        expect(proposalId).toMatch(/^0x[A-Fa-f0-9]{40}_0x[A-Fa-f0-9]{1,64}$/);
 
-        const canExecuteParams: CanExecuteParams = {
-          proposalId,
-          pluginAddress,
-        };
-        let canExecute = await client.methods.canExecute(canExecuteParams);
+        let canExecute = await client.methods.canExecute(proposalId);
         expect(typeof canExecute).toBe("boolean");
         expect(canExecute).toBe(false);
 
         // now approve
-        await voteProposal(pluginAddress, proposalId, client);
+        await voteProposal(proposalId, client);
         // Force date past end
         await mineBlockWithTimeOffset(provider, 2 * 60 * 60);
 
-        canExecute = await client.methods.canExecute(canExecuteParams);
+        canExecute = await client.methods.canExecute(proposalId);
         expect(typeof canExecute).toBe("boolean");
         expect(canExecute).toBe(true);
       });
@@ -324,21 +318,18 @@ describe("Token Voting Client", () => {
         }
 
         const proposalId = await buildProposal(pluginAddress, client);
-        expect(typeof proposalId).toBe("number");
+        expect(typeof proposalId).toBe("string");
+        expect(proposalId).toMatch(/^0x[A-Fa-f0-9]{40}_0x[A-Fa-f0-9]{1,64}$/);
 
-        const canExecuteParams: CanExecuteParams = {
-          proposalId,
-          pluginAddress,
-        };
-        let canExecute = await client.methods.canExecute(canExecuteParams);
+        let canExecute = await client.methods.canExecute(proposalId);
         expect(typeof canExecute).toBe("boolean");
         expect(canExecute).toBe(false);
 
         // now approve
-        await voteProposal(pluginAddress, proposalId, client);
+        await voteProposal(proposalId, client);
         // No waiting
 
-        canExecute = await client.methods.canExecute(canExecuteParams);
+        canExecute = await client.methods.canExecute(proposalId);
         expect(typeof canExecute).toBe("boolean");
         expect(canExecute).toBe(true);
       });
@@ -357,30 +348,27 @@ describe("Token Voting Client", () => {
         }
 
         const proposalId = await buildProposal(pluginAddress, client);
-        expect(typeof proposalId).toBe("number");
+        expect(typeof proposalId).toBe("string");
+        expect(proposalId).toMatch(/^0x[A-Fa-f0-9]{40}_0x[A-Fa-f0-9]{1,64}$/);
 
-        const canExecuteParams: CanExecuteParams = {
-          proposalId,
-          pluginAddress,
-        };
-        let canExecute = await client.methods.canExecute(canExecuteParams);
+        let canExecute = await client.methods.canExecute(proposalId);
         expect(typeof canExecute).toBe("boolean");
         expect(canExecute).toBe(false);
 
         // vote no
-        await voteProposal(pluginAddress, proposalId, client, VoteValues.NO);
+        await voteProposal(proposalId, client, VoteValues.NO);
 
-        canExecute = await client.methods.canExecute(canExecuteParams);
+        canExecute = await client.methods.canExecute(proposalId);
         expect(typeof canExecute).toBe("boolean");
         expect(canExecute).toBe(false);
 
         // now approve
-        await voteProposal(pluginAddress, proposalId, client, VoteValues.YES);
+        await voteProposal(proposalId, client, VoteValues.YES);
 
         // Force date past end
         await mineBlockWithTimeOffset(provider, 2 * 60 * 60);
 
-        canExecute = await client.methods.canExecute(canExecuteParams);
+        canExecute = await client.methods.canExecute(proposalId);
         expect(typeof canExecute).toBe("boolean");
         expect(canExecute).toBe(true);
       });
@@ -401,21 +389,18 @@ describe("Token Voting Client", () => {
         }
 
         const proposalId = await buildProposal(pluginAddress, client);
-        expect(typeof proposalId).toBe("number");
+        expect(typeof proposalId).toBe("string");
+        expect(proposalId).toMatch(/^0x[A-Fa-f0-9]{40}_0x[A-Fa-f0-9]{1,64}$/);
 
         // Vote
-        await voteProposal(pluginAddress, proposalId, client);
+        await voteProposal(proposalId, client);
         // Force date past end
         await mineBlockWithTimeOffset(provider, 2 * 60 * 60);
 
         // Execute
-        const executeParams: IExecuteProposalParams = {
-          pluginAddress,
-          proposalId,
-        };
         for await (
           const step of client.methods.executeProposal(
-            executeParams,
+            proposalId,
           )
         ) {
           switch (step.key) {
@@ -448,20 +433,17 @@ describe("Token Voting Client", () => {
         }
 
         const proposalId = await buildProposal(pluginAddress, client);
-        expect(typeof proposalId).toBe("number");
+        expect(typeof proposalId).toBe("string");
+        expect(proposalId).toMatch(/^0x[A-Fa-f0-9]{40}_0x[A-Fa-f0-9]{1,64}$/);
 
         // Vote
-        await voteProposal(pluginAddress, proposalId, client);
+        await voteProposal(proposalId, client);
         // No waiting here
 
         // Execute
-        const executeParams: IExecuteProposalParams = {
-          pluginAddress,
-          proposalId,
-        };
         for await (
           const step of client.methods.executeProposal(
-            executeParams,
+            proposalId,
           )
         ) {
           switch (step.key) {
@@ -494,23 +476,20 @@ describe("Token Voting Client", () => {
         }
 
         const proposalId = await buildProposal(pluginAddress, client);
-        expect(typeof proposalId).toBe("number");
+        expect(typeof proposalId).toBe("string");
+        expect(proposalId).toMatch(/^0x[A-Fa-f0-9]{40}_0x[A-Fa-f0-9]{1,64}$/);
 
         // Vote
-        await voteProposal(pluginAddress, proposalId, client, VoteValues.NO);
+        await voteProposal(proposalId, client, VoteValues.NO);
         await mineBlock(provider);
-        await voteProposal(pluginAddress, proposalId, client, VoteValues.YES);
+        await voteProposal(proposalId, client, VoteValues.YES);
         // Force date past end
         await mineBlockWithTimeOffset(provider, 2 * 60 * 60);
 
         // Execute
-        const executeParams: IExecuteProposalParams = {
-          pluginAddress,
-          proposalId,
-        };
         for await (
           const step of client.methods.executeProposal(
-            executeParams,
+            proposalId,
           )
         ) {
           switch (step.key) {